+++ packages/ui/src/proxyApi/index.ts
@@ -1,1 +1,1 @@
-export * from './client/api'
+export * from './client/ProxyApi'
